renamed header.js to index, updated import statement, cleaned up code in gatsby-config
