+++ gatsby-config.js
@@ -29,8 +29,8 @@ module.exports = {
     {
       resolve: `gatsby-plugin-manifest`,
       options: {
-        name: `gatsby-starter-default`,
-        short_name: `starter`,
+        name: `Coding Nexus`,
+        short_name: `CN`,
         start_url: `/`,
         background_color: `#663399`,
         theme_color: `#663399`,
