+++ src/layouts/layout.js
@@ -1,14 +1,7 @@
-/**
- * Layout component that queries for data
- * with Gatsby's useStaticQuery component
- *
- * See: https://www.gatsbyjs.com/docs/use-static-query/
- */
-
 import React from 'react';
 import { useStaticQuery, graphql } from 'gatsby';
 
-import Header from '../components/Header/header';
+import Header from '../components/Header/';
 import './layout.css';
 
 const Layout = ({ children }) => {
